+++ core/binary_classification.py
@@ -1,23 +1,27 @@
 import numpy as np
 import pandas as pd
+from sklearn import metrics
 from sklearn.cluster import KMeans
-from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, confusion_matrix
 
-from core.common import choose_classifier
+from core.common import choose_classifier, calculate_results, add_actual
 from core.constants import KMEANS, NO_CLUSTER
 
 pd.options.mode.chained_assignment = None
 
 
-def next_activity(training_df, test_df, job: dict):
+def binary_classifier(training_df, test_df, job):
+    """For True/False classification"""
     clf = choose_classifier(job)
 
-    train_data, test_data, original_test_data = drop_columns(training_df, test_df)
+    training_df, test_df = add_actual(training_df, test_df)
 
+    train_data, test_data, original_test_data = drop_columns(training_df, test_df)
     if job['clustering'] == KMEANS:
         results_df, auc, model_split = kmeans_clustering_train(original_test_data, train_data, clf)
-    else:
+    elif job['clustering'] == NO_CLUSTER:
         results_df, auc, model_split = no_clustering_train(original_test_data, train_data, clf)
+    else:
+        raise ValueError("Unexpected clustering {}".format(job['clustering']))
 
     results = prepare_results(results_df, auc)
     return results, model_split
@@ -26,15 +30,15 @@ def next_activity(training_df, test_df, job: dict):
 def kmeans_clustering_train(original_test_data, train_data, clf):
     estimator = KMeans(n_clusters=3, random_state=21)
     models = dict()
-    estimator.fit(train_data.drop('label', 1))
+    estimator.fit(train_data.drop('actual', 1))
     cluster_lists = {i: train_data.iloc[np.where(estimator.labels_ == i)[0]] for i in range(estimator.n_clusters)}
     for i, cluster_list in cluster_lists.items():
         clustered_train_data = cluster_list
         if clustered_train_data.shape[0] == 0:
             pass
         else:
-            y = clustered_train_data['label']
-            clf.fit(clustered_train_data.drop('label', 1), y)
+            y = clustered_train_data['actual']
+            clf.fit(clustered_train_data.drop('actual', 1), y)
 
             models[i] = clf
     model_split = dict()
@@ -46,14 +50,16 @@ def kmeans_clustering_train(original_test_data, train_data, clf):
 
 
 def kmeans_clustering_test(test_data, clf, estimator, testing=False):
-    drop_list = ['trace_id', 'label'] if testing else ['trace_id']
+    drop_list = ['trace_id', 'actual'] if testing else ['trace_id']
     auc = 0
     counter = 0
+
     test_cluster_lists = {
         i: test_data.iloc[np.where(estimator.predict(test_data.drop(drop_list, 1)) == i)[0]]
         for i in range(estimator.n_clusters)}
     result_data = None
     for i, cluster_list in test_cluster_lists.items():
+        counter += 1
         original_clustered_test_data = cluster_list
         if original_clustered_test_data.shape[0] == 0:
             pass
@@ -61,10 +67,12 @@ def kmeans_clustering_test(test_data, clf, estimator, testing=False):
             clustered_test_data = original_clustered_test_data.drop(drop_list, 1)
 
             prediction = clf[i].predict(clustered_test_data)
+            scores = clf[i].predict_proba(clustered_test_data)
 
             original_clustered_test_data["predicted"] = prediction
             if testing:
-                original_clustered_test_data["actual"] = original_clustered_test_data['label']
+                actual = original_clustered_test_data['actual']
+                auc = calculate_auc(actual, scores, auc)
 
             if result_data is None:
                 result_data = original_clustered_test_data
@@ -79,13 +87,18 @@ def kmeans_clustering_test(test_data, clf, estimator, testing=False):
 
 
 def no_clustering_train(original_test_data, train_data, clf):
-    y = train_data['label']
-    clf.fit(train_data.drop('label', 1), y)
-    actual = original_test_data["label"]
-    original_test_data, scores = no_clustering_test(original_test_data.drop('label', 1), clf)
+    y = train_data['actual']
+
+    clf.fit(train_data.drop('actual', 1), y)
+    actual = original_test_data["actual"]
+    original_test_data, scores = no_clustering_test(original_test_data.drop('actual', 1), clf)
     original_test_data["actual"] = actual
-    # TODO calculate AUC
+
     auc = 0
+    try:
+        auc = metrics.roc_auc_score(actual, scores)
+    except ValueError:
+        pass
     model_split = dict()
     model_split['type'] = NO_CLUSTER
     model_split['model'] = clf
@@ -99,30 +112,28 @@ def no_clustering_test(test_data, clf):
     return test_data, scores
 
 
-def prepare_results(df, auc):
+def prepare_results(df, auc: int):
     actual_ = df['actual'].values
     predicted_ = df['predicted'].values
 
-    row = results_multi_label(actual_, predicted_)
+    row = calculate_results(actual_, predicted_)
     row['auc'] = auc
     return row
 
 
 def drop_columns(training_df, test_df):
-    original_test_df = test_df
-    train_df = training_df.drop('trace_id', 1)
-    test_df = test_df.drop('trace_id', 1)
+    training_df = training_df.drop(['label', 'trace_id'], 1)
+    original_test_df = test_df.drop('label', 1)
+    test_df = test_df.drop(['label', 'trace_id'], 1)
+    return training_df, test_df, original_test_df
 
-    return train_df, test_df, original_test_df
 
-
-def results_multi_label(actual: list, predicted: list):
-    # average is needed as these are multi-label lists
-    # print(classification_report(actual, predicted))
-    acc = accuracy_score(actual, predicted)
-    f1score = f1_score(actual, predicted, average='macro')
-    precision = precision_score(actual, predicted, average='micro')
-    recall = recall_score(actual, predicted, average='macro')
-    # confusion matrix is not binary for easy representation, so removing
-    row = {'f1score': f1score, 'acc': acc, 'precision': precision, 'recall': recall}
-    return row
+def calculate_auc(actual, scores, auc: int):
+    if scores.shape[1] == 1:
+        auc += 0
+    else:
+        try:
+            auc += metrics.roc_auc_score(actual, scores[:, 1])
+        except Exception:
+            pass
+    return auc
